# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API do BrokerHub Angola para CRM e gestão integrada de seguros.
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: dashboard
    description: Resumos operacionais da corretora
  - name: clients
    description: Clientes e tomadores de seguro
  - name: policies
    description: Apólices e renovações
  - name: claims
    description: Sinistros
  - name: activity
    description: Linha de atividade
  - name: tenants
    description: Onboarding de corretoras e sessão
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /me:
    get:
      operationId: getMe
      tags: [tenants]
      summary: Sessão, MFA e corretora do utilizador autenticado
      responses:
        "200":
          description: Current session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Me"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /tenants:
    post:
      operationId: createTenant
      tags: [tenants]
      summary: Registar corretora (onboarding, sem dados de demonstração)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantInput"
      responses:
        "201":
          description: Created tenant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Tenant"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /dashboard/summary:
    get:
      operationId: getDashboardSummary
      tags: [dashboard]
      summary: Resumo operacional da corretora
      responses:
        "200":
          description: Dashboard summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardSummary"
  /activities:
    get:
      operationId: listActivities
      tags: [activity]
      summary: Atividade recente
      parameters:
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Activity"
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: Lista pesquisável de clientes
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Client list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
    post:
      operationId: createClient
      tags: [clients]
      summary: Registar cliente
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientInput"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          $ref: "#/components/responses/BadRequest"
  /clients/{clientId}:
    patch:
      operationId: updateClient
      tags: [clients]
      summary: Atualizar cliente
      parameters:
        - $ref: "#/components/parameters/ClientId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdate"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "404":
          $ref: "#/components/responses/NotFound"
  /policies:
    get:
      operationId: listPolicies
      tags: [policies]
      summary: Lista de apólices
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, renewal, expired, cancelled]
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Policy list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Policy"
    post:
      operationId: createPolicy
      tags: [policies]
      summary: Registar apólice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PolicyInput"
      responses:
        "201":
          description: Created policy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Policy"
  /claims:
    get:
      operationId: listClaims
      tags: [claims]
      summary: Lista de sinistros
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_review, approved, rejected, paid]
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Limit"
      responses:
        "200":
          description: Claim list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Claim"
    post:
      operationId: createClaim
      tags: [claims]
      summary: Participar sinistro
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimInput"
      responses:
        "201":
          description: Created claim
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
components:
  parameters:
    ClientId:
      name: clientId
      in: path
      required: true
      schema:
        type: string
    Search:
      name: search
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: number
        minimum: 1
        maximum: 100
        default: 25
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Unauthenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required: [status]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
          enum:
            [
              UNAUTHENTICATED,
              TENANT_REQUIRED,
              MFA_REQUIRED,
              ORIGIN_DENIED,
              VALIDATION_ERROR,
              CONFLICT,
              NOT_FOUND,
            ]
      required: [error]
    Me:
      type: object
      properties:
        userId:
          type: string
        role:
          type: ["string", "null"]
          enum: [super_admin, broker_master, agent, client, null]
        tenant:
          oneOf:
            - $ref: "#/components/schemas/Tenant"
            - type: "null"
        mfa:
          $ref: "#/components/schemas/MfaStatus"
      required: [userId, role, tenant, mfa]
    MfaStatus:
      type: object
      properties:
        required:
          type: boolean
        satisfied:
          type: boolean
      required: [required, satisfied]
    Tenant:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nif:
          type: string
        timezone:
          type: string
        currency:
          type: string
      required: [id, name, nif, timezone, currency]
    TenantInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        nif:
          type: string
          description: NIF angolano (10 dígitos colectivo ou 9+2 letras+3 particular)
      required: [name, nif]
    DashboardSummary:
      type: object
      properties:
        activeClients:
          type: number
        activePolicies:
          type: number
        openClaims:
          type: number
        renewalsThisMonth:
          type: number
        portfolioValue:
          type: number
        commissionValue:
          type: number
        renewalRate:
          type: number
        monthlyTrend:
          type: array
          items:
            $ref: "#/components/schemas/TrendPoint"
        policyMix:
          type: array
          items:
            $ref: "#/components/schemas/PolicyMix"
      required: [activeClients, activePolicies, openClaims, renewalsThisMonth, portfolioValue, commissionValue, renewalRate, monthlyTrend, policyMix]
    TrendPoint:
      type: object
      properties:
        month:
          type: string
        value:
          type: number
      required: [month, value]
    PolicyMix:
      type: object
      properties:
        type:
          type: string
        count:
          type: number
        percentage:
          type: number
      required: [type, count, percentage]
    Activity:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [client, policy, claim, renewal, payment]
        title:
          type: string
        description:
          type: string
        timestamp:
          type: string
        initials:
          type: string
        tone:
          type: string
          enum: [blue, green, orange, purple, red]
      required: [id, type, title, description, timestamp, initials, tone]
    Client:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        nif:
          type: string
        email:
          type: string
        phone:
          type: string
        type:
          type: string
          enum: [individual, company]
        policiesCount:
          type: number
        totalPremium:
          type: number
        status:
          type: string
          enum: [active, pending, inactive]
        createdAt:
          type: string
      required: [id, name, nif, email, phone, type, policiesCount, totalPremium, status, createdAt]
    ClientInput:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        nif:
          type: string
        email:
          type: string
        phone:
          type: string
        type:
          type: string
          enum: [individual, company]
      required: [name, nif, email, phone, type]
    ClientUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        phone:
          type: string
        status:
          type: string
          enum: [active, pending, inactive]
    Policy:
      type: object
      properties:
        id:
          type: string
        number:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        insurer:
          type: string
        type:
          type: string
          enum: [auto, health, accident, property, life]
        status:
          type: string
          enum: [active, renewal, expired, cancelled]
        premium:
          type: number
        commission:
          type: number
        startDate:
          type: string
        endDate:
          type: string
      required: [id, number, clientId, clientName, insurer, type, status, premium, commission, startDate, endDate]
    PolicyInput:
      type: object
      properties:
        clientId:
          type: string
        insurer:
          type: string
        type:
          type: string
          enum: [auto, health, accident, property, life]
        premium:
          type: number
        commission:
          type: number
        startDate:
          type: string
        endDate:
          type: string
      required: [clientId, insurer, type, premium, commission, startDate, endDate]
    Claim:
      type: object
      properties:
        id:
          type: string
        reference:
          type: string
        clientName:
          type: string
        policyNumber:
          type: string
        type:
          type: string
        insurer:
          type: string
        status:
          type: string
          enum: [open, in_review, approved, rejected, paid]
        amount:
          type: number
        createdAt:
          type: string
        daysOpen:
          type: number
      required: [id, reference, clientName, policyNumber, type, insurer, status, amount, createdAt, daysOpen]
    ClaimInput:
      type: object
      properties:
        policyId:
          type: string
        type:
          type: string
        amount:
          type: number
        description:
          type: string
      required: [policyId, type, amount, description]